+++ Transpilers/Inline/Inline.PHP.psx.ps1
@@ -1,0 +1,45 @@
+<#
+.SYNOPSIS
+    PHP PipeScript Transpiler.
+.DESCRIPTION
+    Transpiles PHP with Inline PipeScript into PHP.
+
+    Multiline comments blocks like this ```<!--{}-->``` will be treated as blocks of PipeScript.
+
+    JavaScript/CSS/PHP comment blocks like ```/*{}*/``` will also be treated as blocks of PipeScript.
+#>
+[ValidateScript({
+    $cmdInfo = $_
+    if ($cmdInfo.Source -match '\.php$') {
+        return $true
+    }
+    return $false    
+})]
+param(
+# The command information.  This will include the path to the file.
+[Parameter(Mandatory,ValueFromPipeline)]
+$CommandInfo
+)
+
+begin {
+    # We start off by declaring a number of regular expressions:
+    $startComment = '(?><\!--|/\*)' # * Start Comments ```<!--```
+    $endComment   = '(?>-->|\*/)'   # * End Comments   ```-->```
+    $Whitespace   = '[\s\n\r]{0,}'
+    # * StartRegex     ```$StartComment + '{' + $Whitespace```
+    $startRegex = "(?<PSStart>${startComment}\{$Whitespace)"
+    # * EndRegex       ```$whitespace + '}' + $EndComment```
+    $endRegex   = "(?<PSEnd>$Whitespace\}${endComment}\s{0,})"
+
+    $sourcePattern  = [Regex]::New("(?>$(
+        $startRegex, $endRegex -join ([Environment]::NewLine + '|' + [Environment]::NewLine)
+    ))", "IgnoreCase, IgnorePatternWhitespace", "00:00:05")
+}
+
+process {
+    
+    $fileInfo = $commandInfo.Source -as [IO.FileInfo]
+    $fileText      = [IO.File]::ReadAllText($fileInfo.Fullname)
+
+    .>PipeScript.Inline -SourceFile $CommandInfo.Source -SourceText $fileText -SourcePattern $sourcePattern
+}